requirements.txt
# Core
torch>=2.3.0
torchvision>=0.18.0
onnx>=1.16.0
onnxruntime>=1.18.0
onnxsim>=0.4.36
onnxscript>=0.1.0
numpy<2.0
pillow>=10.0.0
apache-tvm
jupyter
pytest
